Added music + sound effect capabilities
Added ca.keal.sastrane.gui.audio, which contains utility classes for working with audio. These are Music, which plays music (and can randomly shuffle), and SoundEffects, which can play up to 2 sound effects simultaneously and holds a cache so that the sound file does not have to be loaded every time the effect is played.
Also added a central Random instance - Utils.RANDOM. Added Utils.randomChoice(), which chooses a random element out of an array.