Added dynamic title, Main.STAGE
The window title is now dynamic; this means that the window title will change when you enter into a game to the name of that game followed by a i18nized suffix.
Main saves its primaryStage into Main.STAGE for convenience; this replaces all the GuiUtils.getStage() stuff that was there before.