Added notation infrastructure
Added everything necessary for games to optionally provide notation methods. This is done through implementing Notatable.
Upon the end of a notatable game, a button underneath the "Quit" button (or whatever it's called) says "Notation" and will take the user to a SimpleText page with the output from the game's notate(List<String>) method.
Implementors of Notatable are recommended to be liberal with linebreaks.