Separated notater from notatable
Notatable things now return a Notater, which contains the notation algorithm. The chess notater is now in LongAlgebraicNotater.
The reasoning behind this is to divide up the files more logically. It doesn't make sense to have notation in the main game file -- although this is certainly possible (especially for very simple notation schemes), since Notater is a functional interface.