Made FXML custom components work with Guice
Modified GuiUtils from loadCustom to a guicification routine. Anyways, custom components now work. We would use FXMLComponentBuilderFactory from fx-guice as the builder factory for the FXMLLoaders to guicify, but it has fx-guice's @FXMLComponent hardcoded as the component annotation, and it's a scope annotation; Guice complains when using it with AssistedInject. So we defined our own @FXMLComponent and copied FXMLComponentBuilderFactory using it, then reflectively returned a FXMLComponentBuilder (package-private in fx-guice) as the builder.