Added auto-adding attributes
Generics Hell. Gah. Anyways, changed GameAttrib and AbstractGameModule to allow for *auto-adding attributes*, attributes that add themselves once their dependencies are fulfilled. This is used for the resource bundle name, i18n name, icon resource, and css resource.
configure() must now be called at the *end* of the game module.
Added PossiblyTypedValue<T> as package-private in ca.keal.sastrane.api as an aid to implementing the auto-adding attributes.
Suffered through nine trials of generics hell. Auto-adding attributes are implemented by keeping a map of attributes -> PossiblyTypedValues, but the map loses information or something? Before PossiblyTypedValue was added it was an absolute nightmare.