Added attribute defaults, DEFAULT_REGISTRATOR
Added defaults for attributes. That means that PLACING_PIECES' default is now `new Player[0]`, etc. Defaults are filled in at the very end, so therefore you can override them.
Added the DEFAULTS_REGISTRATOR attribute. In this final preparation for getting rid of Game, DEFAULTS_REGISTRATOR replaces Game.registerDefaults(EventBus). It is a Consumer<EventBus> and defaults to `b -> {}`.
Added a convenience constructor to PossiblyTypedValue taking only a `T value`; it simply defers to its `Class<T>, T` constructor, using `value.getClass()` as the `Class<T>`. Java complains about it, because why wouldn't it?