Removed @Singleton from arbitrators and notaters
There's no reason for the arbitrators/notaters to be singletons. Removed that.
The remaining singletons have a good reason to be: Music, SoundEffects, and GameRegistrar must be singletons by their very nature, and we would construct a *lot* of usages of GuiUtils if it wasn't a singleton - it has no state as well.