Added passing, auto-passing
In a game module, you can enable ALLOW_PASSING to allow players and AIs to pass at any time. Separately, you can enable AUTO_PASS to make the game pass for you if there are no valid moves left.
Added a "pass" button at the bottom of the game screen for when ALLOW_PASSING is enabled. It won't show if both players are AIs.
To prevent getting stuck in a loop in which the AIs perpetually pass, AIs will not pass if the last move was a pass, except if it is the only possible move.