Put pieces on top of diagonal lines in xiangqi
Went back to the old system of putting diagonal lines in the grid, but used a translate(X|Y) value of 1 to align them perfectly. Overhauled the boardGrid in GameController; now there's a StackPane that encompasses the background and the square, and decor is inserted in-between. To set the background in CSS, use the selector `.background` instead of `.square` and set the background on that.