Add Territory, GoUtils
Added Territory, a class *very* similar to Chain that represents a
territory of a player; that is, a group of connected empty points
surrounded by that player's stones. Territory is *very* similar to
Chain; the two should be somehow merged eventually.

Also added GoUtils and moved Stone.getAdjacent to it, as that method
was needed in Territory (operating not on stones, but on squares).

Territory will be used in determining the score at the end of the game.