Add GoArbitrator, bind in GoModule
Added GoArbitrator, the arbitrator for go. Adhering to Chinese rules,
GoArbitrator considers the game over when the last two moves were
passes, and calculates each player's score by adding the number of
that player's stones on the board to the amount of territory (as
determined by Territory.getAllOfPlayer) controlled by that player.

The komi (a bonus given to White to make up for moving second) is
7.5 points under Chinese rules, meaning that there can never be a tie
under these rules. The komi is binded in GoModule and injected into
the arbitrator.