Fix sizing of go's grid
The squares on go's grid now have no padding. Because of this, the
grid appeared very small when the dimensions of the grid itself were
not adjusted. To fix this, used a DEFAULTS_REGISTRATOR in GoModule
to register an anonymous Object subclass containing one method: an
@Subscribe to ToGameEvent.Post which rebinds the grid's width/
height properties to 0.8 of the scene's width/height, instead of 0.6. It
now looks good.

We should probably provide an API for resizing the grid in game
modules, instead of relying on defaults registrators. That should
happen.