Fix go not ending on double pass
A game of go ends when black and white pass consecutively. This was
not happening because Round wasn't putting passes into its `moves`
list (a List<StateChange>). Fixed that, now it does.

However, this broke reversi's GridNotater, which didn't handle passes.
It now does, and simply skips the move (but still increments the move
number!) if it's `== Move.PASS`.

All notaters in games where there is a possibility of passing must be
careful to handle the special case of Move.PASS, as its getEndPos()
returns null, and it is not a MovingMove or a PlacingMove.