Increase performance in PlacingPiece's algorithms
Replaced PlacingPiece.getPossiblePlacements() with PlacingPiece.
getMoveAt(), and had PlacingPiece implement getPossiblePlacements()
based on getMoveAt(). (As a consequence, PlacingPiece is now an
abstract class.) Then, used getMoveAt() in GameController instead of
getPossiblePlacements(). This should increase performance significantly
as we are no longer testing *every* square in order to get the move for
the *one* square we want.