Add some tests for Round, TestUtils
Add some tests for Round. RoundTest contains tests for Round's
constructor, copy constructor, and one for willAutoPass(Player). They
are not complete, and should be at some later time.

TestUtils is a class in a new package, ca.keal.sastrane.test, which
contains useful things for writing tests, such as the TEST_GAMEID
constant. It is used by RoundTest.