Add Game class
~2 months later, I realized how mangled the DI usage is. So, we're
going back to the original method (kinda). One object, a Game class,
which is DI'd into anything that needs info about the game. Kapich?

Each game will provide an implementation of Game, and will bind that in
its module, instead of individual stuff. Eventually, GameAttr and the
like will be completely removed, and it'll be testable. Probably.
Hopefully. In all honesty, probably not.