Add click for AI, delay if AI turn takes < 600ms
If the AI takes less than 600 milliseconds to come up with its next
move, the game now waits for the remainder of the 600ms before
displaying the move. This provides a better user experience -- it's not
fun when the AI's move appears *immediately* after your own.

Also add click when the AI moves, like when a human moves.